Implement the record assembly algorithm
Parquet shreds records to a columnar form and assembles them back using
the encoding described in the Dremel paper:
https://research.google/pubs/pub36632/

In short, adding two small integers (called the repetition and definition levels)
it is possible to efficiently convert JSON-like records (of a know schema)
to columnar form and back. Importantly, this technique also allows reading
back only a subset of the full schema without losing the nested structure.

record_reader implements the assembly algorithm. It's used through a template
parameter, which provides it with callback like "start_struct", "end_list",
"append_value", "separate_list_values", etc.. This interface was chosen with
a record printing use case in mind, and so it easily translates to printing.

This is a recursive implementation. It's performance should be investigated.
Deep call chains for simple tasks, big number of virtual calls and heavy usage
of seastar::repeat and friends in a place of normal loops may prove too slow.

Maybe the algorithm should be written as a state machine (as seen in the paper)
or maybe it's a use case for seastar threads?