Add file_writer
file_writer opens a new Parquet file of a given name and schema.
It spawns and exposes column_chunk_writers for each column,
flushes row_groups on demand and writes the file metadata to output when
the file is being closed.

The chunk writers have to be in sync on every row group flush
(that is, every writer has to receive the same number of fields between flushes).
This is a responsibility of the user - missing values won't be automatically
filled in with nulls or anything of the sort.