Add thrift serialization/deserialization routines
Parquet metadata is serialized with Thrift's TCompactProtocol.

Due to a design error in Parquet (namely, the inclusion of min/max fields
inside the page header), the only way to find out the size of a serialized
page header is to deserialize it. (This has been acknowledged as a mistake by
the authors and metadata correcting this mistake have since been added.
Unfortunately, this fact is useless in the general case, because that metadata
is optional thanks to backwards compatibility).

This blunder means that Parquet can't be comfortably read from a simple byte
stream - we need to progressively grow the stream buffer until the page header
is deserialized without an error, and then we have to store the excessive data
from the buffer somewhere, because it contains the beginning of the next page.

So much for a format designed for sequential IO.

This mess is dealt with by peekable_stream.