Add column_chunk_reader
For parallelization, records in a Parquet file are logically
partitioned into independent groups, called row groups.
(The underlying intention is that each partition can live in a separate file,
on a single Hadoop filesystem block).

A row group is stored on disk in a columnar form, obtained with the
Dremel record shredding algorithm. A chunk of data for a single column in
a row group is called a column chunk and is the basic unit of Parquet
data.

To read Parquet data, the reader should find out from the file metadata
the offset and size of a column chunks that interests him.
A column chunk can then be read from the given location sequentially,
in a streaming fashion.

Column chunks are composed of compressed and encoded units called pages,
which are simply concatenated.

A column_chunk_reader is the core reader interface of this library.
Its job is to read a page from the input stream, decompress it,
initialize decoders with the decompressed data and then extract batches
of values and levels from the decoders for the user.
When the page runs out of values, repeat the cycle, until there is no more
pages.
A column_chunk_reader also owns the column dictionary (shared by all pages
in the chunk).