Add column_chunk_writer
column_chunk_writer is the counterpart of column_chunk_reader.
It buffers the values put in by the user.
When a page flush request comes, the values are grouped into a page,
encoded, compressed, and stored.
When a chunk flush request comes, the stored compressed pages are written to
the output stream. This returns the writer returns to its initial state.

The encoding and compression algorithms used can be specified
at construction time.

The writer can't decide to flush itself on its own,
because flushes have to happen simultaneously for all columns in the row group.
Therefore, the user of this class has to flush the writer periodically,
so that pages and chunks don't grow too big.
To this end, the writer provides methods for estimating the size of
buffered data. (Those methods have a non-zero cost and shouldn't be used
after every single value written, though).