Add support for logical types
In addition to a physical type (int32, byte array), a Parquet column
may be annotated with a logical type (timestamp, utf8, decimal, etc.).

Unfortunately, the authors of Parquet lacked the foresight to design this
annotation well, and added another, better annotation later.
Due to backward compatibility now we have to deal with them both, which makes
the code rather messy.

The rules and interpretation of logical types can be found at:
https://github.com/apache/parquet-format/blob/master/LogicalTypes.md