Add reader schema
Thrift does not support recursive data structures,
so the schema is stored as a list obtained with preorder traversal
of the schema tree.

The code in this commit recovers the tree structure,
and computes the definition and repetition levels of schema elements.
The resulting `raw_schema` provides sufficient information
to start decoding individual column chunks.

To assemble complex records from individual fields, another schema
transformation is needed: the conversion rules for MAP and LIST logical types
have to be applied to the `raw_schema`.
This yields the full `schema`, which will be needed to create a `record_reader`.