Add plain, dictionary and RLE decoders
Before compression, pages of data are encoded. Available encodings are
described in detail at
https://github.com/apache/parquet-format/blob/master/Encodings.md

This commit adds plain, dictionary and RLE decoders.

Plain encoding is trivial - the data are simply written in little endian
and concatenated. Plain encoding is used to encode dictionary pages and
and as fallback for dictionary encoding, when the dictionary page grows
too large.

RLE encoding is an encoding designed for low cardinality integers of known
bit-width. It is used for encoding definition and repetition levels and
for encoding dictionary indices in dictionary encoding.

Dictionary encoding is the main value encoding used in Parquet.
In the dictionary encoding, a column chunk begins with a dictionary page,
which maps integer indices to values. Then, pages of RLE-encoded indices follow.

These are the most important encodings. There other encodings will be added
later, when the general design of the library stabilizes.
We can copy them from Apache Arrow, which contains C++ implementations
of all Parquet encodings (although they are not made available to users for some
reason).