Fix interpretation of empty groups in schema
reader_schema assumed in some places that a node with no children is a
primitive node. That's not the case, because empty group nodes are not
prohibited by Parquet spec. This mistake caused files containing empty group
nodes to be interpreted incorrectly / rejected as invalid.
Fixed.

Interestingly, Apache Arrow currently makes the same mistake.